// src/app/model/BorrowerDocsData.ts
export class BorrowerDocsData {
    idPassport: boolean;
    idPassNumber: string;
    paperPassSeries: string;
    paperPassNumber: string;
    issueDate: any;
    issuer: string;
    region: string;
    district: string;
    postalCode: string;
    settlement: string;
    street: string;
    number: string;
    corpsNo: string;
    door: string;
    itnNumber: string;
    fileName: string;
}